% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Biplot.R
\name{Plot_Biplot}
\alias{Plot_Biplot}
\title{Plotting Biplot}
\usage{
Plot_Biplot(X, axis = c(1,2),
  color = "red", shape = 20, size = 4,
  ind.label = FALSE, ind.label.size = 4,
  arrow.col = "black", vec.name = TRUE, angle.vec = FALSE)
}
\arguments{
\item{X}{List containing the output of one of the functions of the package.}

\item{axis}{Vector with lenght 2 which contains the axis ploted in x and y axis.}

\item{color}{Points colors. It can be a character indicating the color of all the points or a factor to use different colors.}

\item{shape}{Points shape. It can be a number to indicate the shape of all the points or a factor to indicate different shapes.}

\item{size}{numeric value indicating the size of points.}

\item{ind.label}{Logical value, if it is TRUE it prints the name for each row of X. If it is FALSE (default) does not print the names.}

\item{ind.label.size}{numeric value indicating the size of the labels of points.}

\item{arrow.col}{Character indicating the color of the arrows.}

\item{vec.name}{Logical value, if it is TRUE (default) it prints the name for each column of X. If it is FALSE does not print the names.}

\item{angle.vec}{Logical value, if it it TRUE (default) it print the vector names with orentation of the angle of the vector. If it is FALSE the angle of all tags is 0.}
}
\value{
Return a \code{\link{ggplot2}} object.
}
\description{
\code{Plot_Biplot} initializes a ggplot2-based visualization of the caracteristics presented in the data analized by the Biplot selected.
}
\examples{
hj.biplot <- HJBiplot(mtcars, Transform.Data = 'scale')
Plot_Biplot(hj.biplot, ind.label = TRUE)

}
\seealso{
\code{\link{HJBiplot}}, \code{\link{Ridge_HJBiplot}}, \code{\link{ElasticNet_HJBiplot}}
}
\author{
Mitzi Cubilla-Montilla, Carlos Torres-Cubilla, Ana Belen Nieto Librero and Purificacion Galindo Villardon
}
