% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Biplot.R
\name{Plot_Biplot}
\alias{Plot_Biplot}
\title{Plotting Biplot}
\usage{
Plot_Biplot(X, groups = NULL, ind.name = FALSE, vec.name = TRUE,
point.col = "red", arrow.col = "black", axis = c(1,2), angle.vec = TRUE)
}
\arguments{
\item{X}{List containing the output of one of the functions of the package.}

\item{groups}{Factor wich contains groups to consider in plot. NULL groups considered by default.}

\item{ind.name}{Logical value, if it is TRUE it prints the name for each row of X. If it is FALSE (default) does not print the names.}

\item{vec.name}{Logical value, if it is TRUE (default) it prints the name for each column of X. If it is FALSE does not print the names.}

\item{point.col}{Character indicating the color of the points. If groups is not NULL it have to be a vector with the same number of groups factors}

\item{arrow.col}{Character indicating the color of the arrows.}

\item{axis}{Vector with lenght 2 which contains the axis ploted in x and y axis.}

\item{angle.vec}{Logical value, if it it TRUE (default) it print the vector names with orentation of the angle of the vector. If it is FALSE the angle of all tags is 0.}
}
\value{
Return a \code{\link{ggplot2}} object.
}
\description{
\code{Plot_Biplot} initializes a ggplot2-based visualization of the caracteristics presented in the data analized by the Biplot selected.
}
\examples{
hj.biplot <- HJBiplot(mtcars, Transform.Data = 'scale')
Plot_Biplot(hj.biplot)

}
\seealso{
\code{\link{HJBiplot}}, \code{\link{Ridge_HJBiplot}}, \code{\link{ElasticNet_HJBiplot}}
}
\author{
Mitzi Cubilla-Montilla, Carlos Torres-Cubilla, Ana Belen Nieto Librero and Purificacion Galindo Villardon
}
